all TypeProtect tests pass

+++ src/core/file_sys/devices/base_device.h
@@ -9,7 +9,7 @@
 
 namespace Core {
 struct VmObject;
-enum class MemoryProt : u32;
+enum class MemoryProt : u8;
 enum class MemoryMapFlags : u32;
 }
